feat(main/main.js): _getChangedCardProperties() returns model index of changed task card
In pevious implementation, this method scanned for rendered cards and returned the index of the target card. This was based on view, not model, and was vulnerable to bugs. Now _getChangedCardProperties() returns data-task-object-index, which is an actual index of card in the model, thus ensures that the right card was deleted or marked completed.